refactor(application): 상수 및 Controller 시작으로 변경
Before:

scope: {
Application에서 Calculator를 직접 호출
Calculator에서 하드코딩으로 숫자를 입력
}

After:

scope: {
Calculator 내에 사용하는 상수로 변경
Application 클래스는 Controller 로직만 호출하도록 변경
}
